Fix bad memory bug in instruction processing
Previously, the n_operands field of the instruction added to the list
was not always set correctly (specifically, it was not set correctly
for instructions with pseudo-operands). This meant that the
instructions_clear function would sometimes try to free the memory
occupied by an operand which didn't exist. This was not a problem when
I initialized all the operands to NULL (since freeing NULL is a
no-op), but arose when I tried to take this initialization out.