Remove 'numtest' test
The code in the test was taken from a program (BLINKY) whose license I'm not
totally sure about, and also it was intended to test something that has since
been fixed.  Some new assembly files to test the assembler/disassembler should
be coming soon, so its loss shouldn't affect much.