Use callback system for drawing
Instead of redrawing the entire screen and then refreshing the window on
each draw, this commit introduces a draw callback.  Each time a pixel
needs to be drawn (turned on or off), the callback is called with the
coordinates and the new value.  Thus, only the areas that are modified
are redrawn, although the entire window is updated (using
SDL_UpdateWindowSurface) after every draw.

This *significantly* improves speed and memory usage; profiling revealed
that the old system spent at least around 80-90% of its time in the
drawing function.