Implement sleep-based draw timing
Instead of waiting for a timer latch, where the timer is constantly
increasing based on the system clock, the draw instructions are now
delayed by an actual nanosleep(2) call.  This results in a lot less CPU
usage, since the OS can do other things with the CPU while the emulator
is sleeping.