Fix handling of non-quirky shifts
As it turns out, the reason why ANT didn't work before was that the
shift instructions SHL and SHR, when interpreted without quirks, were
too strict: they required vy to be 0 in the opcode. This, however,
conflicted with the ANT code, which has vy as other values, even though
it expects the non-quirky behavior of ignoring vy.